blocks some features for demo security

+++ app/Actions/Jetstream/DeleteUser.php
@@ -18,7 +18,7 @@ class DeleteUser implements DeletesUsers
     /**
      * Create a new action instance.
      *
-     * @param  \Laravel\Jetstream\Contracts\DeletesTeams  $deletesTeams
+     * @param \Laravel\Jetstream\Contracts\DeletesTeams $deletesTeams
      * @return void
      */
     public function __construct(DeletesTeams $deletesTeams)
@@ -29,23 +29,25 @@ class DeleteUser implements DeletesUsers
     /**
      * Delete the given user.
      *
-     * @param  mixed  $user
+     * @param mixed $user
      * @return void
      */
     public function delete($user)
     {
-        DB::transaction(function () use ($user) {
-            $this->deleteTeams($user);
-            $user->deleteProfilePhoto();
-            $user->tokens->each->delete();
-            $user->delete();
-        });
+        if ($user->id !== 1) {
+            DB::transaction(function () use ($user) {
+                $this->deleteTeams($user);
+                $user->deleteProfilePhoto();
+                $user->tokens->each->delete();
+                $user->delete();
+            });
+        }
     }
 
     /**
      * Delete the teams and team associations attached to the user.
      *
-     * @param  mixed  $user
+     * @param mixed $user
      * @return void
      */
     protected function deleteTeams($user)
